Comment out an error message added in the previous patch.
The error message reported what looks like an error condition:
the config file doesn't provide transport info.  However what
appears to be happening is that the program tries to read
several different sections of the config file, of which some
are relevant and some aren't, due to some overloading of the
<routers> tag.  As a result it routinely spawns irrelevant
children, only to watch them die.

There's got to be a better way, but for now I'll just
suppress the error message.

M    src/router/osrf_router_main.c


git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@1800 9efc2488-bf62-4759-914b-345cdb29e865