Eliminated two members of the osrfMessage structure: result_string and sender_tz_offset.  Neither is used for anything, and either may be easily reinstated if necessary.
I was tempted to eliminate the protocol member as well,
since it isn't used for anything either.  However it's
populated from one of the parameters to osrf_message_init().
Getting rid of the protocol member properly would ramify
to all the code that calls osrf_message_init(), which would
be a lot of work to undo if necessary, so I left it alone.

M    include/opensrf/osrf_message.h
M    src/libopensrf/osrf_message.c


git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@1856 9efc2488-bf62-4759-914b-345cdb29e865