Miscellaneous minor changes. mostly for clarity:
1. Changed return type of _osrfAppRegisterSysMethods from int to void,
since we never looked at the return code anyway.

2. In osrfAppRegisterApplication(): 0pen the shared object before
allocating an osrfApplication, so that we don't have to free the
osrfApplication if the open fails.

3. In osrfAppRegisterExtendedMethod(): when creating an atomic method,
pass the userData pointer to _osrfAppBuildMethod() instead of
installing it on a separate line.

4. In _osrfAppBuildMethod(): for an atomic method, build the method
name correctly the first time, instead of building it incorrectly
and later replacing it.

5. In osrfAppRunMethod(): rearranged things a bit for clarity.
Simplified the declaration and dereferencing of the meth pointer.

6. In osrfAppIntrospect(): Introduced an early return in order to
reduce the level of indentation of the rest of the function.  Moved
some declarations closer to their first uses.

7. In osrfAppIntrospectAll(): moved the declaration of resp into the
loop where it is used.

8. Finished adding doxygen-style comments to document the functions.
Touched up the white space here and there.

M    src/libopensrf/osrf_application.c


git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@1935 9efc2488-bf62-4759-914b-345cdb29e865