Rearrange the way that a listener juggles its drones.
Instead of maintaining a single circular linked list including both active
processes and idle processes, keep the active drones in the circular list
and the idle drones in a separate linear list; move them back and forth
as needed.  This change simplifies how we search for drones when we need
them, because we don't have to skip over ones that we don't want.

The list of idle drones acts as a stack.  When we look for an idle drone
for a new request, we pick the one at the head of the list, which is the
last one to have become idle.

As a result, we pick the drone that is most likely still to be in memory.

In the old arrangment, the list acted as a queue.  When picking an idle
drone, we picked the one that had been idle the longest -- which was the
one most likely to have been swapped out.

Also: added a number of doxygen-style comments, and tinkered with the
white space here and there.

Note: an earlier commit tried to make similar changes, but it had to be
backed out because it mangled the linkage pointers in some situations,
leading to segfaults.  That problem has now been corrected.

M    src/libopensrf/osrf_prefork.c


git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@1967 9efc2488-bf62-4759-914b-345cdb29e865